removed TNamed inheritence and changed vector(float) to double
git-svn-id: https://root.cern.ch/svn/root/branches/dev@26339 27541ba8-7e3a-0410-8455-c3a389f83636

+++ roostats/roofit/roostats/inc/HybridCalculator.h
@@ -50,7 +50,7 @@ namespace RooStats {
       void PrintMore(const char* options);
 
    private:
-      void RunToys(std::vector<float>& bVals, std::vector<float>& sbVals, unsigned int nToys, bool usePriors);
+      void RunToys(std::vector<double>& bVals, std::vector<double>& sbVals, unsigned int nToys, bool usePriors);
 
       RooAbsPdf& fSbModel;
       RooAbsPdf& fBModel;
